Pushing changes to handle transfer using pessimistic locking git status

+++ multithreading/src/main/java/multithreading/bankingApplication/operations/AccountOperationsUsingOptimisticLocking.java
@@ -7,7 +7,7 @@ import multithreading.bankingApplication.exception.WithdrawException;
 
 public class AccountOperationsUsingOptimisticLocking {
 
-    public double fetchBalance(Account account) {
+    public int fetchBalance(Account account) {
         return account.getAtomicBalance().get();
     }
